Bugfix: Add used field to MD5 hash state
MD5.jl currently uses undocumented and unexported internals of the SHA stdlib.
PR #86 to SHA.jl added a new field, `used` to the SHA states. Since MD5 uses
internal SHA functions which now look for this field, MD5 tests began to fail.

This PR adds the needed `used` field, so MD5 now runs. The real fix is to not
rely on SHA internals.